requirements.txt
plotly==5.18.0
pandas==1.5.2
streamlit==1.29.0
streamlit-folium==0.18.0
haversine==2.8.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.24.0
folium==0.15.1
Pillow==9.0.1